+++ app/src/main/java/me/romanow/lep500/fft/FFTAudioSource.java
@@ -14,7 +14,7 @@ import java.io.IOException;
 public interface FFTAudioSource {
     public String testSource(int sizeHZ);
     public long getFrameLength();
-    public int read(float buf[], int offset, int lnt) throws IOException;
+    public int read(double buf[], int offset, int lnt) throws IOException;
     public void close() throws IOException;
     public void enableToPlay(boolean play);
     public void play(int start, int delay);
